Show the subresults and their checks in the HTML report
Fix the colspan of the main table, add a new column "Actions" into the
table and add possibility to show/hide the checks and subresults row for
each result.

Also, show all the checks in the subresults table. Do not hide them by
default as it's done with general result checks.